requirements.txt
# ─────────────────────────────────────────────────────────────────────
# ChatDEVA v2 — Production Requirements
# Optimized for Railway deployment (CPU only, fast build)
# torch is installed separately in nixpacks.toml (CPU wheel)
# ─────────────────────────────────────────────────────────────────────

# ── Web framework ─────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Frontend ──────────────────────────────────────────────────────────
streamlit==1.35.0
requests==2.31.0

# ── Auth ──────────────────────────────────────────────────────────────
bcrypt==4.1.3
python-jose[cryptography]==3.3.0

# ── Database ──────────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9        # PostgreSQL driver (Railway Postgres)

# ── LangChain (pinned — avoids import breaking changes) ───────────────
langchain==0.1.20
langchain-community==0.1.20
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langchain-huggingface==0.0.3

# ── HuggingFace ───────────────────────────────────────────────────────
transformers==4.40.0
huggingface-hub==0.23.0
tokenizers==0.19.1
accelerate==0.29.3
sentence-transformers==3.0.0

# ── Vector store ──────────────────────────────────────────────────────
chromadb==0.5.0

# ── Document loaders ──────────────────────────────────────────────────
pypdf==4.2.0
docx2txt==0.8

# ── Config ────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Numerical ─────────────────────────────────────────────────────────
numpy==1.26.4

# NOTE: torch is installed via nixpacks.toml with CPU-only wheel
# to avoid the 2GB+ CUDA download on Railway
